Document new config file options.
This commit adds documentation for the `auto_restart`, `fast_inject`,
and `secure_input_notification` config file options. It also
alphabetizes the sections of the relevant tables, and fixes a few typos.

It does not attempt to substantially alter the documentation structure.
The current table-based format isn't great for conveying more detailed
information, such as the rationale behind options or their use cases.

For now, I've just jammed brief descriptions of the new options into the
existing table. I'll try to rework the page in the future when I have
more time.

Addresses, but does not fully resolve, issue #357.